fix: band-aid fix to disallow linux video fullscreen
Ref #1022
