+++ lib/common/view/audio_tile_option_button.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:watch_it/watch_it.dart';
 import 'package:yaru/yaru.dart';
 
+import '../../app_config.dart';
 import '../../extensions/build_context_x.dart';
 import '../../l10n/l10n.dart';
 import '../../library/library_model.dart';
@@ -10,7 +11,6 @@ import '../data/audio.dart';
 import 'icons.dart';
 import 'snackbars.dart';
 import 'stream_provider_share_button.dart';
-import 'theme.dart';
 
 class AudioTileOptionButton extends StatelessWidget {
   const AudioTileOptionButton({
