+++ lib/common/view/progress.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:yaru/yaru.dart';
 
+import '../../app_config.dart';
 import '../../extensions/build_context_x.dart';
 import 'theme.dart';
