+++ lib/common/view/search_button.dart
@@ -1,7 +1,9 @@
-import '../../extensions/build_context_x.dart';
-import 'icons.dart';
 import 'package:flutter/material.dart';
 import 'package:yaru/yaru.dart';
+
+import '../../app_config.dart';
+import '../../extensions/build_context_x.dart';
+import 'icons.dart';
 import 'theme.dart';
 
 class SearchButton extends StatelessWidget {
