+++ lib/player/view/player_track.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:watch_it/watch_it.dart';
 
+import '../../app_config.dart';
 import '../../common/data/audio.dart';
 import '../../common/view/custom_track_shape.dart';
 import '../../common/view/progress.dart';
-import '../../common/view/theme.dart';
 import '../../extensions/build_context_x.dart';
 import '../../extensions/duration_x.dart';
 import '../player_model.dart';
